rename slice -> board-slice
