+++ src/scenes/modal/modals/help.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import { useAppDispatch } from '../../../app/hooks';
-import { setGameStatus } from '../../../app/slice';
+import { setGameStatus } from '../../../app/board-slice';
 import { Button } from '../../../components/button';
 import { getColor } from '../../../themes';
